backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
SQLAlchemy>=1.4
asyncpg>=0.27.0
pydantic[email]>=1.10
python-dotenv>=1.0.0
colorlog
yoyo-migrations>=8.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.3.0
python-multipart>=0.0.9